Potentially fixed clipping issue by setting the sin wave table to have a maximum value of 255 (2^8 - 1). Builds, not tested.

+++ main/waveform_tables.h
@@ -1,8 +1,8 @@
 #ifndef WAVEFORM_TABLES_H_
 #define WAVEFORM_TABLES_H_
 
-extern const uint32_t sin_array[256];
-extern const uint32_t sawtooth_array[256];
+extern const uint_fast16_t sin_array[256];
+extern const uint_fast16_t sawtooth_array[256];
 extern const uint32_t ratio_num[128];
 extern const uint32_t ratio_denom[128];
